% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_undefined.R
\name{add_undefined}
\alias{add_undefined}
\title{Add a column specifying if the values in the segmentation are undefined}
\usage{
add_undefined(data, undefined)
}
\description{
Adds a column `is_defined` which indicates if a particular A-scan could be
reliably segmented. NOTE: This function assumes there is a column named
"value", but this column  name is likely to change in future iterations of
the software.
}
