% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_identity.R
\name{anonymize_identity}
\alias{anonymize_identity}
\title{Anonymize identity in a volume object.}
\usage{
anonymize_identity(header, pid, patient_id = pid)
}
\arguments{
\item{header}{The header list from a volume object.}

\item{anon_pid}{A new internal patient ID.}

\item{anon_patient_id}{A new patient ID.}
}
\value{
A new header list with the ID values changed.
}
\description{
\code{anonymize_identity} changes the 'pid' (an internal identifier used by
the Heidelberg database) and 'patient_id' (typically a hospital number, e.g.,
medical record number or MRN) values in a
volume object header. NOTE: By default the patient's name is part of the
file name, so if you are writing out the volume object data to a file,
remember to rename the file accordingly.
}
\keyword{internal}
