% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_image_id.R
\name{anonymize_image_id}
\alias{anonymize_image_id}
\title{Anonymize the image ID in a volume object.}
\usage{
anonymize_image_id(header)
}
\arguments{
\item{header}{The header list from a volume object.}
}
\value{
A new header list with the image ID and reference IDs values changed.
}
\description{
\code{anonymize_image_id} changes the 'id' (an internal identifier used by
the Heidelberg database) and 'reference_id' (if present) values in a
volume object header.
}
\keyword{internal}
