% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_volume.R
\name{anonymize_volume}
\alias{anonymize_volume}
\title{Anonymize volume objects.}
\usage{
anonymize_volume(volume, pid, patient_id = pid,
  anon_dob = as.POSIXct(0, origin = "1899-12-30", tz = "UTC"))
}
\arguments{
\item{volume}{A volume object}

\item{anon_dob}{The new date of birth. Defaults to the origin in Microsoft's
DATE specification as used by the 'dob' and 'visit_date' values in the
Heidelberg VOL file specification (HSF-OCT-101).}

\item{anon_pid}{A new internal patient ID.}

\item{anon_patient_id}{A new patient ID.}
}
\value{
A new volume object with the identifiers and dates adjusted.
}
\description{
\code{anonymize_volume} updates the patient ID, date of birth, and date
fields specified in Heidelberg's HSF-OCT-101 file format specification.
NOTE: This function will not catch identifying data if it is included in the
'spare' bytes or other fields in the volume object. We strongly recommend
that you inspect your own VOL files for other potentially identifying
fields not documented in HSF-OCT-101.
}
