% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_bscan.R
\name{construct_bscan}
\alias{construct_bscan}
\title{Construct a ggplot2 object containing the data for a given b-scan.}
\usage{
construct_bscan(oct, bscan_id, layer_y_max = oct$header$size_z,
  layer_y_min = 0, low_color = "black", high_color = "white",
  contrast_correction = spline_correction, na_intensity = 0,
  scale_bars = TRUE, scale_length = 200, scale_color = "white",
  inset_percentage = c(x = 0.025, y = 0.05))
}
\arguments{
\item{oct}{A list containing the OCT data.}

\item{bscan_id}{The number of the b-scan of interest.}

\item{layer_y_max}{Vertical maximum constraint on the b-scan axix.}

\item{layer_y_min}{Vertical minimum constraint on the b-scan axis.}

\item{low_color}{Color for lowest values.}

\item{high_color}{Color for highest values.}

\item{scale_bars}{Boolean indicating if scale bars be drawn.}

\item{scale_length}{Length of scale bars in microns.}

\item{scale_color}{Color of scale bar.}

\item{inset_percentage}{Percentage of image to inset the scale bar.}
}
\value{
A ggplot2 object.
}
\description{
\code{construct_bscan} builds a ggplot2 object containing the data for a
given b-scan.
}
