% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_slo.R
\name{construct_slo}
\alias{construct_slo}
\title{Construct a ggplot2 object containing the data for an SLO}
\usage{
construct_slo(oct, low_color = "black", high_color = "white",
  scale_bar = TRUE, scale_length = 200, scale_color = "white",
  inset_percentage = 0.05, draw_margins = TRUE)
}
\arguments{
\item{oct}{a list containing the OCT data}

\item{low_color}{color for lowest values}

\item{high_color}{color for highest values}

\item{scale_length}{length of scale bars in microns}

\item{scale_color}{color of scale bar}

\item{inset_percentage}{percentage of image to inset the scale bar}

\item{draw_margins}{Should the plot include margins and axis labels?}

\item{scale_bars}{Should scale bars be drawn?}
}
\value{
a ggplot2 object
}
\description{
Construct a ggplot2 object containing the data for an SLO
}
