% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{datetime_to_raw}
\alias{datetime_to_raw}
\title{datetime_to_raw}
\usage{
datetime_to_raw(sec_epoch)
}
\arguments{
\item{x}{Integer representing the seconds from the epoch}
}
\value{
a raw vector of 8 bytes representing the number of 100 nanosecond
units since 1601-01-01.
}
\description{
Covert a datetime value, encoded as an integer, to the binary representation
of Windows filetime to POSIXct.
}
\details{
\code{raw_to_datetime} takes an raw vector encoding a 64-bit unsigned integer
This value is the number of 100 nanoseconds since 1601-01-01.
}
