% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delist_singletons.R
\name{delist_singletons}
\alias{delist_singletons}
\title{Remove singleton lists of a list.}
\usage{
delist_singletons(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A list.
}
\description{
\code{delist_singletons} searches a list for elements that are list of length
1 and "delists" by assigning the contents to the name of the list.
}
\examples{
mylist <- list("A" = 1, "B" = list(1:5), "C" = list(little = 1:3, tiny = 4:5))
mylist
delist_singletons(mylist)
}
