% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_mapping.R
\name{downsample_mapping}
\alias{downsample_mapping}
\title{Downsample a sequence of numbers.}
\usage{
downsample_mapping(original_n, down_n)
}
\arguments{
\item{original_n}{An integer giving the length of the original sequence of
numbers.}

\item{down_n}{The desired size of the downsampled sequence.}
}
\value{
A list containing the downsampling factor and a
    vector a downsampled indices.
}
\description{
\code{downsample_mapping} downsamples a sequence of numbers, 1 through
\code{n} and returns a vector of indices, centered on the middle
number of the original sequence.
}
