% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_vol.R
\name{downsample_vol}
\alias{downsample_vol}
\title{Downsample B-scans in a volume object.}
\usage{
downsample_vol(vol, new_n)
}
\arguments{
\item{vol}{A volume object.}

\item{new_n}{The number of B-scans to retain.}
}
\value{
A volume object.
}
\description{
\code{downsample_vol} reduces the number of B-scans in a volume
object. The reduction maintains the central B-scan, and ensures that
the distance between B-scans is constant.
}
