% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_segmentation_xml.R
\name{filter_segmentation_xml}
\alias{filter_segmentation_xml}
\title{Filter surfaces from an XML segmentation file}
\usage{
filter_segmentation_xml(
  xml_file,
  keep_surfaces = c(10, 20, 30, 40, 50, 60, 100, 110, 120, 140, 150),
  modifier = "heyexr"
)
}
\arguments{
\item{xml_file}{An path or connection to an XML file or a character
representation of XML.}

\item{keep_surfaces}{A vector of the surface IDs to keep in the file. Use
the labels specified by OCTExplorer. By default all surfaces segmented by
IRA/OCTExplorer ver. 3.8.0 are retained.}
}
\value{
The filtered XML object.
}
\description{
Filters surfaces from an XML segmentation to reduce the number of surfaces.
}
