% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_region.R
\name{flip_region}
\alias{flip_region}
\title{Flip grid region for a table of thickness values}
\usage{
flip_region(x, flip_to = "OD")
}
\arguments{
\item{x}{a data.frame of thickness values. Must contain 'grid', 'laterality', and 'region' columns.}

\item{flip_to}{laterality to enforce for all measurements}
}
\value{
a tbl_df of the original data plus two new columns, 'flipped_laterality' and 'flipped_region'
}
\description{
Ensures that all thickness values are comparable across regions
}
