% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bscan.R
\name{get_bscan}
\alias{get_bscan}
\title{Parse the b-scan data from an OCT list object}
\usage{
get_bscan(oct, n = 1)
}
\arguments{
\item{oct}{the OCT list object}

\item{n}{the ID of the requested b-scan}
}
\value{
a tbl_df containing the b-scan
}
\description{
Parse the b-scan data from an OCT list object
}
