% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_central_segmentation.R
\name{get_central_segmentation}
\alias{get_central_segmentation}
\title{Get the segmentation data from the central scan}
\usage{
get_central_segmentation(
  vol_file = NULL,
  xml_file = NULL,
  center_file = NULL,
  sample_id = NA
)
}
\arguments{
\item{xml_file}{path to XML file containing Iowa Reference Algorithms segmentation}

\item{center_file}{path to XML file specifying the foveal center of the VOL data}
}
\value{
a data.frame (tbl_df) containing the central segmentation data
}
\description{
Retrieves the segmentation data for the central b-scan, as specified in an
XML file.
}
