% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segmentation.R
\name{get_segmentation}
\alias{get_segmentation}
\title{Get the segmentation data from an OCT object}
\usage{
get_segmentation(oct)
}
\arguments{
\item{oct}{OCT list object}
}
\value{
a tbl_df of the segmentation data
}
\description{
Retrieve the automated Heidelberg segmentation data from an OCT object
}
