% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slo.R
\name{get_slo}
\alias{get_slo}
\title{Get the SLO data from an OCT object}
\usage{
get_slo(oct)
}
\arguments{
\item{oct}{an OCT list oct}
}
\value{
a tbl_df of the SLO data
}
\description{
Retrieves the SLO data from an OCT object
}
