% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{grid_regions}
\alias{grid_regions}
\title{Corresponding grid regions for right (OD) and left (OS) eyes.}
\format{
A tbl_df with 9 rows and 3 variables:
\describe{
  \item{grid}{grid, the underlying XML file defining the grid used by OCT Explorer to measure thickness}
  \item{laterality}{laterality, anatomic position of the eye}
  \item{region}{region, region in the grid used to measure thickness}
  \item{flipped_laterality}{flipped_laterality, transformed anatomic position of the eye}
  \item{flipped_region}{flipped_region, transformed region in the grid used to measure thickness}
}
}
\usage{
grid_regions
}
\description{
A dataset containing a mapping between grid regions in OS and OD eyes.
}
\keyword{datasets}
