% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_description.R
\docType{package}
\name{heyexr}
\alias{heyexr}
\alias{_PACKAGE}
\alias{heyexr-package}
\title{heyexr: Import Heidelberg Spectralis OCT VOL Files}
\description{
Import and  Heidelberg Spectralis optical coherence tomography (OCT) data in R.
}
\author{
S. Scott Whitmore <steven-whitmore@uiowa.edu>
}
