% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_thickness_csv.R
\name{import_thickness_csv}
\alias{import_thickness_csv}
\title{Import the CSV produced by "Thickness Analysis" in OCT Explorer}
\usage{
import_thickness_csv(
  csv_file,
  split = "\\\\\\\\",
  n = "last",
  pattern = "_Surfaces_Iowa.xml"
)
}
\arguments{
\item{csv_file}{the CSV file to import}

\item{split}{a character string to use as the split pattern}

\item{n}{either "last" or an integer position}

\item{pattern}{character string to remove from the values in the Surfaces column}

\item{scrub}{a string to remove from the file column generated by OCT Explorer}
}
\description{
Imports and formats the CSV data produced by the "Thickness Analysis"
performed by OCT Explorer software. OCT Explorer is currently distributed for
Windows, and the "Surfaces" column contains the full path to each XML
segmentation file, producing unwieldy identifiers. Therefore, handy default
values for the parameters split, n, and pattern are provided for cleaning up
the values in the Surfaces column.
}
