% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{layer_info}
\alias{layer_info}
\title{Metadata on retinal layers identified by the Iowa Reference Algorithms}
\format{A tibble}
\usage{
layer_info
}
\description{
Metadata on segmentation surfaces and layers identified by the Iowa
Reference Algorithms (v 3.8.0).
}
\keyword{datasets}
