% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_plot_2.R
\name{layout_plot_2}
\alias{layout_plot_2}
\title{Layout plots}
\usage{
layout_plot_2(bscan_id, oct, p_slo, layer_y_max, layer_y_min, xml_file,
  oct_segmentation, bscan_id_seg, center_file, center_z_voxel,
  center_x_voxel, overlay_bscan_position, overlay_heidelberg_segmentation,
  oct_seg_array, base_name, output_path, file_type, tt)
}
\value{
a list containing the OCT and segmentation data (if requested)
}
\description{
UPDATE
}
