% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_onerow.R
\name{list_to_onerow}
\alias{list_to_onerow}
\title{Convert a list to a single-row tibble.}
\usage{
list_to_onerow(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A tibble.
}
\description{
\code{list_to_onerow} converts lists with multiple levels to a data frame of
 a single row.
}
\examples{
mylist <- list("A" = 1, "B" = c(1,2,4),
               "C" = letters[1:5],
               "D" = list(10:20, fruit = c("orange", "apple", "banana")))
mylist
list_to_onerow(mylist)
}
