% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{max_float}
\alias{max_float}
\title{Value used to indicate missing data (NA).}
\format{
A scalar double.
}
\usage{
max_float
}
\description{
Maximum floating point value used by Heidelberg Engineering to represent
missing data.
}
\keyword{datasets}
