% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_raw.R
\name{pad_raw}
\alias{pad_raw}
\title{\code{pad_raw} adds null bytes to the end of a raw vector to give a vector of
a specified size.}
\usage{
pad_raw(x, n, size = 1L)
}
\arguments{
\item{x}{A raw vector.}

\item{n}{The total number of elements in the output vector.}

\item{size}{The number of bytes taken up by each element.}
}
\value{
A raw vector of length n.
}
\description{
\code{pad_raw} adds null bytes to the end of a raw vector to give a vector of
a specified size.
}
