% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote_elements.R
\name{promote_elements}
\alias{promote_elements}
\title{Convert a list to a wide tibble.}
\usage{
promote_elements(x, combine_names = TRUE, recursive = FALSE)
}
\arguments{
\item{x}{A list.}
}
\value{
If any of the elements of the list had multiple value, these
  values will now be included as separate elements in the top level.
}
\description{
\code{promote_elements} converts multi-value elements of a list into uniquely
named elements of the [primary] list. [What's a better way to say this?]
}
\examples{

mylist <- list("A" = 1, "B" = c(1,2,4),
               "C" = letters[1:5],
               "D" = list(10:20, fruit = c("orange", "apple", "banana")))
mylist
promote_elements(mylist)

# The following two lines are equivalent for this example:
promote_elements(promote_elements(mylist))
promote_elements(mylist, recursive = TRUE)
}
