% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raw_to_datetime}
\alias{raw_to_datetime}
\title{raw_to_datetime}
\usage{
raw_to_datetime(src)
}
\arguments{
\item{x}{a raw vector of 8 bytes representing the number of 100 nanosecond
units since 1601-01-01.}
}
\description{
Covert the binary representation of Windows filetime to POSIXct
}
\details{
\code{raw_to_datetime} takes an raw vector encoding a 64-bit unsigned integer
This value is the number of 100 nanoseconds since 1601-01-01.
}
