% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bscan_header.R
\name{read_bscan_header}
\alias{read_bscan_header}
\title{Read b-scan header information from a connection to a VOL file}
\usage{
read_bscan_header(vol_con)
}
\arguments{
\item{vol_con}{the connection to a VOL file}
}
\value{
a list of the b-scan specfific header information
}
\description{
Reads the b-scan header information from a connection to a VOL file. Function
based on Open_Heyex_Raw.java from the HEYEX plugin for ImageJ.
}
