% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_center_xml.R
\name{read_center_xml}
\alias{read_center_xml}
\title{Read "grid center" XML file generated by the Iowa Reference Algorithms}
\usage{
read_center_xml(center_file)
}
\arguments{
\item{center_file}{path to the XML file containing the grid center coordinates}
}
\value{
a list containing the x and y coordinates of the grid center
}
\description{
Reads the coordinates of the grid center as output by the Iowa Reference
Algorithms (XML format).
}
