% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_float.R
\name{read_float}
\alias{read_float}
\title{Read a float from a binary connection}
\usage{
read_float(con)
}
\arguments{
\item{con}{a file connection}
}
\value{
a numeric
}
\description{
Reads 4 bytes in little Endian and converts to a float
}
