% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_float_vector.R
\name{read_float_vector}
\alias{read_float_vector}
\title{Read an vector of floats from a binary connection}
\usage{
read_float_vector(con, n = 1)
}
\arguments{
\item{con}{a file connection}

\item{n}{size of vector}
}
\value{
a numeric
}
\description{
Reads n number of 4 bytes in little Endian and converts numeric vector
}
