% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_oct_files.R
\name{read_oct_files}
\alias{read_oct_files}
\title{Read files related to one OCT image}
\usage{
read_oct_files(volume = NA, segmentation = NA, grid_center = NA,
  disk_center = NA)
}
\description{
Reads a set of files produced by OCTExplorer related to a single OCT image.
}
