% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ssoct_files.R
\name{read_ssoct_files}
\alias{read_ssoct_files}
\title{Read OCT volume, angiography, segmentation, and grid center.}
\usage{
read_ssoct_files(
  volume = NA,
  angio = NA,
  segmentation = NA,
  grid_center = NA,
  disk_center = NA,
  laterality = NA
)
}
\value{
A list of objects.
}
\description{
Read OCT volume, angiography, segmentation, and grid center.
}
