% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vol.R
\name{read_vol}
\alias{read_vol}
\title{Read a Heidelberg Spectralis VOL file}
\usage{
read_vol(vol_file, read_what = "all")
}
\arguments{
\item{vol_file}{Path to VOL file.}

\item{read_what}{Character string indicating what parts of the file to read.
If "all", then all data is read. If "slo", then only the header and SLO
are read. Any other value will only return the header.}
}
\value{
a list containing the data from the VOL file
}
\description{
Creates a list containing data from a Heidelberg Spectralis VOL file.
}
