% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vol_header.R
\name{read_vol_header}
\alias{read_vol_header}
\title{Read the VOL header}
\usage{
read_vol_header(vol_con, tz = "UTC")
}
\arguments{
\item{vol_con}{a connection to the VOL file}

\item{tz}{Timezone for dates in header.}
}
\value{
a list containing the header from the VOL file
}
\description{
Read the header information from a Heidelberg Spectralis VOL file.
Function assumes the header begins at byte 0. Due to limitation in R,
this function does not currently parse date and time information correctly.
}
