% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vol_noloop.R
\name{read_vol_noloop}
\alias{read_vol_noloop}
\title{Read a Heidelberg Spectralis VOL file}
\usage{
read_vol_noloop(vol_file, header_slo_only = FALSE)
}
\arguments{
\item{vol_file}{path to VOL file}

\item{header_slo_only}{Import only the header information and SLO image?}
}
\value{
a list containing the data from the VOL file
}
\description{
Creates a list containing data from a Heidelberg Spectralis VOL file.
}
