% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vol_slo.R
\name{read_vol_slo}
\alias{read_vol_slo}
\title{Read the VOL SLO image}
\usage{
read_vol_slo(vol_con, header)
}
\arguments{
\item{header}{previously read header information}

\item{con}{connection to a VOL file}
}
\value{
a matrix containing the SLO image
}
\description{
Read the SLO data from a Heidelberg Spectralis VOL file.
}
