% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_oct_summary.R
\name{render_oct_summary}
\alias{render_oct_summary}
\title{Render summary plots of SLO and b-scans for an OCT VOL file}
\usage{
render_oct_summary(vol_file, xml_file = NULL, center_file = NULL,
  out_dir = "rendered_bscans", return_results = FALSE, prefix = "",
  overlay_bscan_position = TRUE, omit_slo = FALSE,
  overlay_heidelberg_segmentation = TRUE,
  crop_to_heidelberg_segmentation = c(10, 75), file_type = "pdf",
  slo_only = FALSE)
}
\arguments{
\item{vol_file}{path to VOL file}

\item{xml_file}{path to XML file containing Iowa Reference Algorithms segmentation}

\item{out_dir}{name of directory to save rendered results}

\item{return_results}{Return the data read as a list?}
}
\value{
a list containing the OCT and segmentation data (if requested)
}
\description{
Generates plots from a VOL file. Segmentation from the Iowa Reference
Algorithms can be included as an optional overlay on the b-scans.
}
