% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_slo.R
\name{render_slo}
\alias{render_slo}
\title{Render SLO data}
\usage{
render_slo(vol_file, out_dir = ".", draw_margins = FALSE,
  type = "png", size = 4, units = "in", ...)
}
\arguments{
\item{out_dir}{directory to save image}

\item{type}{image format (device)}

\item{size}{x/y dimensions in units}

\item{units}{units for x/y dimensions}

\item{file}{a VOL file}
}
\value{
none
}
\description{
Reads SLO data from a VOL file and renders as an image in the specified
format.
}
