% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_correction.R
\name{spline_correction}
\alias{spline_correction}
\title{Correct contrast}
\usage{
spline_correction(i)
}
\value{
A vector of transformed intensities
}
\description{
Adjust the intensity values using a spline model
based on the difference contrast correction performed
by Heidelberg Explorer.
}
