% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_nth.R
\name{strsplit_nth}
\alias{strsplit_nth}
\title{Split a string and retain the nth value}
\usage{
strsplit_nth(x, split, n = "last", ...)
}
\arguments{
\item{x}{a character vector}

\item{split}{a character string to use as the split pattern}

\item{n}{either "last" or an integer position}

\item{...}{parameters to pass to strsplit}
}
\value{
a tbl_df of the segmentation data
}
\description{
Retrieve the automated Heidelberg segmentation data from an OCT object
}
