% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nude.R
\name{theme_nude}
\alias{theme_nude}
\title{Render a ggplot2 plot with only the data.}
\usage{
theme_nude(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\description{
\code{theme_nude} themes a ggplot2 plot with only the data. Remember to remove any padding by
also adding `+ scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0))`.
This function is useful when overlaying raster data on en face and B-scan
images.
}
