% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data.R
\docType{data}
\name{vol_fit_spline}
\alias{vol_fit_spline}
\title{A spline fit to the PNG exported from Heidelberg Explorer.}
\format{
A smooth.spline object.
}
\usage{
vol_fit_spline
}
\description{
A smooth.spline object.
}
\keyword{datasets}
