% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bscan_header.R
\name{write_bscan_header}
\alias{write_bscan_header}
\title{Write b-scan header information to a file connection}
\usage{
write_bscan_header(vol_con, bscan_header_list)
}
\arguments{
\item{vol_con}{the connection to a VOL file}

\item{bscan_header_list}{previously imported b-scan header data}
}
\description{
Writes the b-scan header information to a file connection.
}
