% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vol.R
\name{write_vol}
\alias{write_vol}
\title{Write the data in a volume object to a VOL file.}
\usage{
write_vol(volume, vol_file, overwrite = FALSE)
}
\arguments{
\item{volume}{A volume object.}

\item{vol_file}{The target file.}

\item{overwrite}{If the target exists, should it be overwritten?}
}
\description{
\code{write_vol} writes a volume object to a VOL file.
TASK: DESCRIBE DIFFERENCES BETWEEN THE HSF-OCT-101 FORMAT AND WHAT THIS
FUNCTION PRODUCES.
}
