% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vol_header.R
\name{write_vol_header}
\alias{write_vol_header}
\title{Write a VOL header to a file connection.}
\usage{
write_vol_header(vol_con, header)
}
\arguments{
\item{vol_con}{File connection object.}

\item{header}{Header (list) from a VOL object.}
}
\description{
\code{write_vol_header} writes the data in a VOL object to a file connection.
}
