% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vol_slo.R
\name{write_vol_slo}
\alias{write_vol_slo}
\title{Write the VOL SLO image}
\usage{
write_vol_slo(vol_con, slo_image)
}
\arguments{
\item{con}{connection to a VOL file}

\item{slo}{SLO data matrix}
}
\description{
Writes the SLO information from a Heidelberg Spectralis VOL file.
}
